// src/core/app/enums/contract_status.ts
export enum ContractStatus {
  INCOMPLETE = "incomplete",
  REQUESTED = "requested",
  WAITING_MARKETER = "waiting-marketer",
  WAITING_CLIENT_SIGN = "waiting-client-sign",
  SIGNED = "signed",
  ACTIVATED = "activated",
  FINISHED = "finished",
  MARKETER_ISSUE = "marketer-issue",
  DISTRIBUTOR_ISSUE = "distributor-issue",
  CANCELLED = "cancelled"
}